Rename lnd/service.ts -> lnd/index.ts
